Fix liquidity_consumption default and sort_price_raw type coercion
- engine.rs: Pass Some(true) for liquidity_consumption to enable L2
  orderbook consumption instead of relying on default
- matching_core: Use i128::from(p.raw) for explicit type coercion in
  sort_price_raw to avoid implicit conversion issues

+++ crates/execution/src/matching_core/mod.rs
@@ -77,9 +77,9 @@ impl OrderMatchInfo {
     #[must_use]
     pub fn sort_price_raw(&self) -> i128 {
         if let Some(p) = self.limit_price {
-            p.raw
+            i128::from(p.raw)
         } else if let Some(p) = self.trigger_price {
-            p.raw
+            i128::from(p.raw)
         } else {
             i128::MAX
         }
